Remove unused imeDownButton in taskbar
This button was originally forked off from the main back button in [1],
prior to the gesture navigation work, and it was the only back button
shown when the IME was visible.

Starting from [2], we reverted to always show the main back button
regardless of the IME visibility, and to also show the imeDownButton
when not in three button navigation (also submitted prior to the gesture
navigation work). Eventually in [3] the IME started supporting gesture
navigation with its own fork of navigation bar, which shows this button.

Later in [4], the imeDownButton was changed to only be visible in
gesture navigation mode, while the IME can draw its own navigation
buttons (specified through a syspro, defaults to true). Thus it could
only be shown when neither in gesture navigation nor in three button
navigation. However, with the navigation bar taskbar unification, this
button is essentially ignored, as it is not manually added.

In any case, as this is a fork of the main back button, all its
behaviour could (and should) be handled by the main implementation.

  [1]: I4d0ecd0bee0c519892c63eeefef45055b26d349b
  [2]: I7959d26986e546596e2436b8b5eb7668f44b55cf
  [3]: I3e7e1f83554444131e2765dc159617bb9e2337c7
  [4]: Iab8c47521ab9e561881e4ffb45142888e4c427e4

Flag: EXEMPT refactor
Bug: 366129400
Test: verify with IME shown and hidden in 3 button nav and gesture nav,
  in launcher and setup wizard
Change-Id: If337d198ddb0e0e2173167add9991f97988a3a5c