Fix app pair crash
* With the change made in ag/31615582, we assumed
that top level + stage level split roots would always be
included in the transition, that isn't the case when an app pair
is already showing and then we switch to another app pair (stage
roots aren't included since they aren't modified in that scenario)
* Now shell sends the entire tree. Also modified SplitScreenUtils
to consider TRANSIT_CHANGE changes as well for the scenario where
one app of an app pair is already full screen and then we launch an
app pair which has that same app, the stage root there undergoes a change
and not a open/close transition.
* One KI is that the animation doesn't fully work when launching a second
app pair from taskbar when split is already showing (though it is an
improvement from current behavior/animation). This is because when the second
app pair launches, the tasks' snapshot windows/leashes are passed back as
leafs, however these snapshots have not been parented to be under the stage roots.
So both snapshots are positioned at 0, 0 causing overlap. Will fix that in
a followup

Bug: 399037701
Test: Open app pairs from desktop windowing + normal taskbar
Flag: EXEMPT bugfix

Change-Id: I83adffd16fea98b8e877dfbe1ba1ef4a3d08fcb9