Add debounce to TaskView emissions
The emissions of each TaskView on Overview open are usually:
1. task exists
2. icon loaded
3. then thumbnail loaded

If the 3 updates above come in quick succession, then ideally the first 2 will be debounced and the last one will be consumed immediately.

Debouncing only when thumbnails are not present gives the best results. This is because thumbnail loading is a decent predictor of there being no more emissions to come as they are typically the last emission for a TaskView.

Bug: 395294954
Flag: com.android.launcher3.enable_refactor_task_thumbnail
Test: ABTD run of failing metric
Change-Id: Ib57d6dc7d9f58b7bb37ec1177b23d89360c24c0e