PageIndicatorDot Placement && Arrow Touchboxes.
In the previous image tests, the page indicator placement within folders
was inconsistent depending on the number of pages in a folder. This change fixes that by removing the margin_End for the page indicator in folders, as well as fixing the onMeasure call in PageIndicatorDots to be the same as the legacy calculation, just with additional space for the arrows.

Previously, the touchboxes were very small, and difficult to use. To enlarge the touchbox, enlarged padding is used. This means when the indicator & arrows are visible, we need to remove the margin from the layout container and move it to padding for the indicator.

Bug: 394355070
Test: Tested folders with 2, 3, and 18 pages in both left to right
and right to left mode.
Flag: com.android.launcher3.enable_launcher_visual_refresh

Change-Id: I3a5d9d6d77dec963865adede1c4bfe38890d608f