desktop-exploded-view: Fix glitch with thumbnailView.outlineBounds
This is a follow up to ag/32333172. With that change, task thumbnails
are immediately placed in the exploded position and then transformed
into the intermediate positions/sizes that is controlled by the explode
progress.

This CL fixes a glitch where thumbnailView.outlineBounds was not
computed properly. The bounds are now computed per frame (since the task
may animate in from a position that is partially off-screen). The
intersection with the screen bounds rectangle is now done with the
per-frame position of the task window and then crucially scaled to
account for the transformation that is applied to the thumbnail itself.

Flag: com.android.launcher3.enable_desktop_exploded_view
Test: Manual
BUG=399822967

Change-Id: I66d6d9f662f0e70e86ea1befcec4c43d2434fb35