Fix chip elevation when expanded
Update the chip z-index of IconAppChipView when bringing it to front or sending it back. This change fixes the elevation when chip is expanded over another chip for split tasks.

Update the task z-index when bottom row chip is expanded to guarantee the chip will be displayed above other TaskVIews.

Fix: 406696829
Fix: 317007147
Flag: com.android.launcher3.enable_overview_icon_menu
Test: OverviewImageTest
Test: OverviewSplitTaskImageTest
Change-Id: Ie68898beea0a012ba083fe1b950e82896ea14c41