Ensure the back button is visible in SUW for simple view
Simple view is meant to provide an easier setup experience and this
means that the back button must be visible in SUW, even if the new
design choice is applied. This change adds a caveat to enable the back
button for this situation.

Fix: 396559615
Test: Enable/Disable simple view with the new theme applied and observe
the back button appear/disappear
Flag: EXEMPT bugfix

Change-Id: Ic58b5d87b2d0670c468222098a53293ae27b0206