Don't translate the back button if home button is showing for lockscreen
The back button would get translated to the edge of the screen in
lockscreen apps. This was fine but now that the home button is shown in
these apps, they would overlap. This change prevents the back button
from translating to the edge of the screen when the home button is shown
in certain screens when the device is locked.

Fix: 406104663
Test: Enable 3 button nav and open lock screen apps/ emergency page on
large screen device. Observe that the buttons don't overlap anymore.
Flag: EXEMPT bugfix

Change-Id: Ibb963c1a62de80a990f223797d37a0a5ee294a90