When updating workspace items, make sure to not process Folder icons as Workspace items.
Also, in Folder.java the DeviceProfile obtained using the activity,
in this case the Activity is TaskbarActivityContext is returning
a different DeviceProfile form the Launcher which causes the update.

So this triggers an update when is not needed but even if we make
a faulty update the workspace should be able to differentiate
between workspace item and folder items.

This issue was triggered by this flag
com.android.window.flags.enable_taskbar_recents_layout_transition .

Fix: 407630150
Test: NA
Flag: EXEMPT bug fix
Change-Id: I255052506e7d198b1384c15ebfe2f6f2753da1ed