Remove AllAppsState history override.
This was introduced with the Rocket Gesture but causes a weird
transition when quickly going from an app to All Apps and invoking
Back, because it thought the previous state was BACKGROUND_APP.

The default just treats the previous state as NORMAL, which is
correct for All Apps. The OVERVIEW case was also an even older
path when you could access All Apps from Overview.

Bug: Can't find it right now.
Test: Open app, quickly go home and to all apps, invoke back.
Flag: EXEMPT bugfix
Change-Id: I76cad48abdba9b621e80d2f0cc0414b59fa35725