Prewarm renderer for All Apps blur.
We already notify renderer when a non-0 blur is about to be
applied, but this change gives additional heads up. On my device,
I saw the following latency improvements (for sending the alert,
not necessarily when it was acted upon):

 - All Apps swipe transtion: 10ms early
 - Taskbar All Apps button: 30ms early

Also clean up the SurfaceControl.Transactions using
try-with-resources.

Flag: com.android.launcher3.all_apps_blur
Bug: 410973453
Bug: 411292920
Bug: 411033774
Test: Log inspections; will run perf tests with ABTD
Change-Id: I640720767de71f439ba479f83db00c00eefe07dd