Check showHomeBehindDesktop to allow swiping up for overview gesture
With the showHomeBehindDesktop feature enabled, the swipe up from the bottom of the screen inadvertently sets LauncherResumedThroughShellTransition as true. This is because the home container is always running behind freeform tasks and so container.isResumed() is always true. Therefore, add a check for showHomeBehindDesktop to make sure LauncherResumedThroughShellTransition is set to false while the home screen is always shown in the background.
This ensures using the correct InputConsumer to allow swiping into overview.

Bug: 375644149
Flag: com.android.window.flags.show_home_behind_desktop
Test: With freeform tasks shown, swipe up from taskbar to go to overview
Change-Id: I68f45092ddf6ecdbc1d3b259f16048fdc40d8df2