Pass bubble bar bounds to shell on handshake
If the systemui process restarts, as it does as part of tests,
it does not get notified about the bubble bar bounds until they
actually change. This could result in invalid bounds for the
expanded view which would prevent bubbling tasks.

This change ensures that the bounds are sent to shell after the
initial handshake.

20x runs: https://android-build.googleplex.com/builds/abtd/run/L46000030011522666

Bug: 411307439
Test: atest BubbleAnythingTest
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I2a5898f32506529b8860274cdee1fcea1a264e83