Store bubble bar state in sysui proxy
SysUIProxy may not be initialized when we set the launcher bubble bar
state and the listener.
Listener is stored in a field and set in sysui when the proxy
initializion is complete.
Add the same for the launcher bubble bar state. Otherwise bubble bar
state in sysui may not be immediately updated.

Bug: 357115534
Test: restart device, trigger a bubble before device is rotated
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I5dca49981c45b0e4c8c2535a283b9d9241eb47ba