Fix widget size in wallpaper preview when switching the grid
Currently there is a bit of inconsistency in the way grid change
happens.

1. Grid migration decides different size, inflate item uses
different, and scale is applied using yet another size.

2. also when
going from small -> medium we reload data, but if while wallpaper picker
is open, if we revert to small, it leads to a wrong state due to
inconsistencies in #1 and shows cropped widget content. Instead in this
cl, we always reload on grid update, so widgets are shown for correct
size. There is still possibility of issue if the widget didn't respond
with data - but even in public build today there are tons of issues with
widgets on changing grid.

Bug: 408934352
Flag: EXEMPT BUGFIX
Test: See demo.
Change-Id: I91741d762cbf34ab7f9e0269ea307d2aad9e0eda