Return to BackgroundAppState when tapping gesture region in desktop
If the end target is null, that means the gesture did not move
enough to calculate an actual end target, and LAST_TASK is the
default for that case. We probably don't need the desktop check
specifically (could clean it up in a follow up) but since this
bug is specifically for Desktop mode I left it in for now.

Fixes: 359879541
Test: manual, tap Taskbar in Desktop and ensure no flicker
Flag: EXEMPT bugfix

Change-Id: Icc3e645819a22ed96d07a575cdccda30a5d800c1