Make sure animation is cancelled for existing flyout
When we show a new flyout, it is possible for the previous flyout to
still be there and animating out.
We were removing the existing flyout, but did not cancel the animation.
This meant that there could have been a hide animation running for the
previous flyout. Once that animation finished, it removed the current
flyout from the container. It was possible for the current flyout to
already be updated to the new one. In which case the animation for the
previous flyout hid the new flyout.

Bug: 414808503
Test: atest PlatformScenarioTests:BubbleBarTest
Flag: com.android.wm.shell.enable_bubble_bar
Change-Id: I4229297fa999d815c12eefd9d583bdfc915ec00a