Update RecentsView to allow re-apply load plan
If the recents view is not attached to window when reloadIfNeeded is called, then mTaskListChangeId will be set, but applyLoadPlan will exit early without actually applying the loaded tasks. This is an issue since the recents list will remain outdated until the task stack changes. This has been causing issues in tapl tests with recents window enabled. Delaying the setting of mAppliedTaskListChangeId until the end of applyLoadPlan to fix this.

Flag: com.android.launcher3.enable_launcher_overview_in_window
Bug: 377678992
Test: TaplTestsQuickstep.testDismissLastGridRow
Change-Id: Iaa8441a74b54f86167a7963b0005c044f355d94e