Start SurfaceFlinger early wakeup sooner for blur.
Now early wakeup is controlled by BaseDepthController (with the
exception of Taskbar which uses TaskbarOverlayController). It does
all of the following:
 - Increase GPU clock
 - Increase CPU clock (notifyRendererOfExpensiveFrame)
 - Tell SurfaceFlinger to wake up early

By default these happen the first time we render a non-0 blur, but
in some cases we can proactively activate it even earlier:
 - Swipe up to All Apps begins
 - Home screen QSB is tapped

These early cases were already handled for CPU and GPU, but now SF
is also in the loop.

In addition, SF early wakeup is disabled when Launcher is no longer
in the foreground.

Bug: 410973453
Test: Log manually to see early wakeup start/end
Flag: com.android.launcher3.all_apps_blur

Change-Id: I39ac0baf8c706b20f0818d31dbc8165b798ea54a