Fix launching 2nd empty desk
Currently, there is a case that create 1 non-empty desk and
two empty desks, then launch the 2nd empty desk without activating the
1st empty desk, it will get stuck at waitForThumbnailsToUpdate.

The reason is the thumbnailData of the tasks that are checked
in waitForThumbnailsToUpdate has been reset to null.

This fix updates waitForThumbnailsToUpdate by firstly filtering the
tasks based on if they are still visible, and only wait on visible tasks.

Bug: 416040522
Flag: EXEMPT bug fix
Test: mauanl create empty desks and launching
Change-Id: I8741a6abe73511bd2b00f2fdbf375228f798ae21