Prevent unnecessary DisplayInfo refreshment
This CL avoids refreshing DisplayInfo when the display is not freeform.
As notifyConfigChange() is expensive, we want to minimize the call.

Note: enterDesktopByDefaultOnFreeformDisplay only checks the flag state
but showLockedTaskbarOnHome can check if the display is freeform in
addition to the flag state.

Flag: com.android.window.flags.enter_desktop_by_default_on_freeform_displays
Test: TaplTestsLockedTaskbar
Bug: 416253799

Change-Id: Ief298c05c8ce46ed70055139e5e817b41f6a572d