Remove haptics for long presses in 3-button-nav
This CL removes haptics for long presses on 3-button-nav buttons, except
when the long press triggers a user visible action.

This means that for example the haptics for long pressing the recents
button are removed, whereas haptics for long pressing the home button
remain active since that launches circle-to-search.

Bug: 408968370
Test: Manual, i.e. verified that no haptic feedback is played when
      long pressing the recents button
Flag: EXEMPT bugfix
Change-Id: Icfeade8a659a8aa4882ef139696c66f410980303