Clear bubble bar saved states if bubble bar controllers not present.
Added logic to clear bubble bar related portion of the shared state to
prevent bubble bar being expanded after device unfolded back. If sysui
restores the expanded view, there is no harm in clearing the saved
state, but if it fails to restore, we would at least have the bubble bar
in the not-expanded state.

Bug: 416297086
Test: Manual. Create any bubble and expanded when device is unfolded.
Fold the device and unfold it back. Observe expanded bubble is closed.
Test: Manual. Be folded, expand bubbles unfold bubbles remain expanded
Flag: EXEMPT - bug fix

Change-Id: I13448c48137cc944ec506bdd48ec630634ed120c