Fix bar state when overflow is selected initially
Fix the state of the bubble bar when the initial state received
from sysui has the overflow selected, which can happen after
unfolding. This ensures that the arrow position is set correctly,
and that when tapping on the overflow in this state the bar
collapses.

Bug: 392893178
Flag: com.android.wm.shell.enable_bubble_bar
Test: manual
       - have floating bubble and expand
       - selecte overflow
       - unfold
       - observe overflow is expanded and bubble bar arrow position
         is correct
       - tap on the overflow bubble
       - observe bar collapses
Change-Id: I4a3d217ce951c7c7f2d17da89b0cd5905faf0eff