Fix Hotseat flashing during DW to home transition
Problem: upon exiting the deesktop mode we were getting change in recents tasks which we were commiting to ui and that resulted in hotseat flashing.

Solution: since layout transition is only in use while we are in desktop windowing mode, we will only need to change icon layout bound while we are in desktop mode and not when we are exiting.

Test: Manual, Presubmit
Bug: 417696108
Bug: 415700104
Flag: EXEMPT bugfix
Change-Id: Ibcf65df49e7b48312a27d2750d3d0c6c5d392006