Navigate Home to hide tasks on WidgetPicker show
When the widget picker is shown while the show_home_behind_desktop
feature is enabled, it becomes possible for the widget picker to be
opened while freeform tasks are occluding the home screen. To fix this
case, navigate Home to move it in front of occluding tasks, and then
open the widget picker. Navigating home is an asynchronous action, and
widget picker will already be opening once the intent executes, so
exlude closing the widget picker when closing all AbstractFloatingViews
in this case.

This change is intended for the desktop form factor.

Bug: 416057590
Flag: com.android.window.flags.show_home_behind_desktop
Test: Open freeform tasks, open WidgetPicker via menu, make sure that
tasks are hidden and WidgetPicker shows.

Change-Id: I19ebeeb706ae7374df8038d956d453708bc96f7e