desktop-exploded-view: Fix regression in quick switch
This is a stop gap solution to fix an issue when quick-switching to a
desktop. The way layout exploded layout currently happens is that we
immediately lay windows out in their exploded (non overlapping) position
and interpolate to the overlapping default position using an animated
transform. When the original window is larger than its exploded layout
(which is going to be the common case), then the header (which is sized
for the exploded layout) is going to look large when the window is
transformed into its overlapping position.

This is a quick fix that basically just hides the header when it
shouldn't be visible.

Flag: com.android.launcher3.enable_desktop_exploded_view
Test: manual
BUG=b/414991065

Change-Id: I906e7f3cb28c57fc81e64b224ef82554ef4c9a5c