KeyEvent.KEYCODE_TAB changes TaskView
This CL restores previous tabbing behaviour to select each TaskView, rather than enter the TaskView and cycle the children as this behaviour hasn't been fully implemented and there are some issues with live tiles, so restoring the previous behaviour.

Note: rather than revert ag/32253797, I decided to leave it as if the menu is open, then the tabbing behaviour works in the menu, so didn't see a reason to revert it.

Bug: 366177160
Bug: 403292849
Flag: com.android.launcher3.enable_overview_icon_menu
Test: Manual using keyboard navigation
Change-Id: I27660bdac60ee572cd633bd58465a6b8bc66aad2