Handle focus events for Overview live tile
In liveTile mode, an InputConsumer is used for gesture navigation and
this consumes all InputEvent's. As a result, the framework doesn't see
these events and so focus isn't handled.

This CL override's dispatchKeyEvent in some of the key views in
RecentsView and handles changing focus correctly.

Bug: 366177160
Fix: 403292849
Flag: com.android.launcher3.enable_overview_icon_menu
Test: Manual using keyboard navigation
Change-Id: Icc15abb19f72d28340ae944aa59b67ddec2eceda